@@ -13,7 +13,7 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3",
+        "php": ">=5.6",
         "guzzle/guzzle": "3.9.*"
     },
     "require-dev": {
